# Adding some basic perf tests for Vector<T>