@fanyang-mono 2022-01-20
A recent library code change to `System.Numerics.Vector` has caused regressions in Mono runtime (https://github.com/dotnet/runtime/pull/53450). Had we had these microbenchmarks running in the perf lab already, it would have saved me quit some time on tracing the root causes. Without them, I had to work on a much larger test (System.Tests.Perf_String.IndexOfAny).

I hope that this change could be merged soon.